+++ StockApi/Models/DataProviders/Stocks/IndexParam.cs
@@ -5,8 +5,8 @@ namespace StockApi.Models.DataProviders.Stocks;
 internal  struct IndexParam(AbstractPagingRequest req) : IKey
 {
     private string BaseKey { get; set; } = req.KeyWithPrefix();
-    public long PageIndex { get; set; } = req.RequestedPage;
-    public long PageSize { get; set; } = req.RecordsPerPage;
+    public int PageIndex { get; set; } = req.RequestedPage;
+    public int PageSize { get; set; } = req.RecordsPerPage;
     public string Category { get; set; } = "TAIEX";
 
     public string KeyWithPrefix()
